Nightmares for days
Dafuq is markdown... who invented this? WHY did they do it? 

It is literally impossible to left-align an image and add a clean bullet-point list to the right of it. This is the only, very hacky, solution for it. 

Man, when I started this tiny edit, I expected to be done in 2 minutes.. now it cost me over a f***ing hour.


REEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEee